% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arpr-package.R
\docType{package}
\name{arpr-package}
\alias{arpr}
\alias{arpr-package}
\title{arpr: Advanced R Pipes}
\description{
Provides convenience functions for programming with 'magrittr'
    pipes. Conditional pipes, a string prefixer and a function to pipe the
    given object into a specific argument given by character name are
    currently supported. It is named after the dadaist Hans Arp, a friend
    of Rene Magritte.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statnmap/arpr}
  \item Report bugs at \url{https://github.com/statnmap/arpr/issues}
}

}
\author{
\strong{Maintainer}: Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Authors:
\itemize{
  \item Jirka Lewandowski \email{jirka.lewandowski@wzb.eu}
}

}
\keyword{internal}
