% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arpr-package.R
\docType{package}
\name{arpr}
\alias{arpr}
\alias{arpr-package}
\title{Advanced R pipes}
\description{
arpr R package
}
\details{
Advanced R pipes
\tabular{ll}{
Package: \tab arpr\cr
Type: \tab Package\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\references{
\url{http://blog.points-of-interest.cc/}
}
\author{
Jirka Lewandowski \email{jirka.lewandowski@wzb.eu}
}
