% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{const}
\alias{const}
\title{Create a constant function}
\usage{
const(val = NULL)
}
\arguments{
\item{val}{return value of constant function (defaults to NULL)}
}
\value{
A function always returning \code{val} accepting arbitrary arguments (dots)
}
\description{
Create a constant function
}
