% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_extensions.R
\name{browse_r}
\alias{browse_r}
\title{browser() in a magrittr pipe}
\usage{
browse_r(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{passed on to browser()}
}
\value{
Used for side effect. Open a browser inside the pipe workflow.
}
\description{
browser() in a magrittr pipe
}
